+++ spec/services/gcloud_spec.rb
