+++ spec/services/payment_methods/paymongo_spec.rb
@@ -2,7 +2,7 @@ require 'rails_helper'
 
 # https://developers.paymongo.com/docs/testing#basic-test-card-numbers
 
-RSpec.describe Paymongo do
+RSpec.describe PaymentMethods::Paymongo do
   describe 'Payment Intents (Invalid Plans)' do
     it 'raises runtime error for invalid plans' do
       expect { described_class.new_payment_intent(:not_a_plan) }.to raise_error(RuntimeError)
